Create in memory MessageQueue
This will be used for our signalling endpoint later on. Eventually we
need to use some kind of in memory database like Redis or DynamoDB
instead of using a javascript Map, since that doesn't work so well with
the serverless model. This should work for testing locally, and working
on the other WebRTC stuff, we can worry about that later.

The idea behind the MessageQueue more generally is that you can send
messages to a user, and then the user can retreive them later, in a FIFO
mailbox style way.

We can use this along with an endpoint for sending a message, and an
endpoint for polling.