Add an endpoint to create and retrieve messages
The first step is modifying the interface of the MessageQueue to be
asynchronous, really we should have done this before, since when we're
interacting with some database, these operations are going to be
asynchronous, since we're going over the network.

Then we have a function to get the instance of the message queue. We
just use a variable containing the instance. This seems to work
correctly locally, and will probably work on serverless initially, as
long as you're in the same region, and not ending up using multiple
execution contexts to serverless functions.

POST to /api/messages/[id] will create a message for id

GET /api/messages/[id] will get all the pending messages for id