Create a Room abstraction
The main purpose of this is to simplify the interface quite a bit, and
also to make it possible to call multiple clients at once, although I
haven't tested that aspect yet.

The way the abstraction works is that you create a room by either
joining an existing room, or hosting one. The only difference is that
when joining a room you send an initial message to the host.

The protocol works like this:
  1. Send a message to the host
  2. The host tells you all of the IDs to contact (themself included)
  3. You send a message to each ID, telling them you want to call them
     You make sure to create a Peer to be ready to receive their WebRTC
     call
  4. They initiate a WebRTC call with you.

This means you have O(n^2) connections, but this will probably be fine
for a demo anyways.